restructured dirs and rewired routes
