+++ src/main.jsx
@@ -1,0 +1,34 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import HomePage from "./components/HomePage/HomePage.jsx";
+import DestinationPage from "./components/DestinationPage/DestinationPage.jsx";
+import AboutUsPage from "./components/AboutUsPage/AboutUsPage.jsx";
+import NasaCollaboration from "./components/NasaCollaborationPage/NasaCollaborationPage.jsx";
+import "./main.css";
+
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <HomePage />,
+  },
+  {
+    path: "/destination",
+    element: <DestinationPage />,
+  },
+  {
+    path: "/about",
+    element: <AboutUsPage />,
+  },
+  {
+    path: "/collaboration",
+    element: <NasaCollaboration />,
+  },
+
+]);
+
+ReactDOM.createRoot(document.getElementById("root")).render(
+  <React.StrictMode>
+    <RouterProvider router={router} />
+  </React.StrictMode>,
+);
